+++ app/components/omnibox/controllers/point-controller.js
@@ -209,28 +209,6 @@ angular.module('lizard-nxt')
     /**
      * @function
      * @memberOf app.pointCtrl
-     * @description Get correct icon for structure
-     */
-    $scope.getIconClass = function (str) {
-      switch (str) {
-      case 'overflow':
-        return 'icon-overflow';
-      case 'pumpstation':
-        return 'icon-pumpstation-diesel';
-      case 'bridge':
-        return 'icon-bridge';
-      case 'bridge-draw':
-        return 'icon-bridge';
-      case 'bridge-fixed':
-        return 'icon-bridge';
-      default:
-        return 'icon-' + str;
-      }
-    };
-
-    /**
-     * @function
-     * @memberOf app.pointCtrl
      * @description Toggling the view on the table for structure attributes;
      *              Either show the first 3 attributes, OR show all of them
      */